airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
google-cloud-storage
google-cloud-bigquery
apache-airflow-providers-google

